ml-backend/requirements_gpu.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
opencv-python
numpy
ultralytics
transformers
accelerate
bitsandbytes
imagehash
exifread
qdrant-client
python-dotenv
pydantic
requests
